pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fp-llm-otherfucker"
version = "0.1.0"
description = "FPLLMotherfucker: data→predict→optimize→report"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "YOUR_NAME" }]

dependencies = [
    "typer>=0.12",
    "pydantic>=2.7",
    "requests>=2.32",
    "pandas>=2.2",
    "pyarrow>=16",
    "PuLP>=2.9",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "ruff>=0.5",
    "black>=24.4",
    "isort>=5.13",
    "mypy>=1.10",
    "types-requests>=2.32.0.20240622",
    "responses>=0.25",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = { find = { where = ["src"] } }

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"scripts/fetch_fpl.py" = ["B008"]
"scripts/build_features.py" = ["B008"]

[tool.black]
line-length = 100
target-version = ["py311"]
exclude = '''
/(
  \.fpllm
  | \.venv
  | venv
  | build
  | dist
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.ruff_cache
)/
'''

[tool.isort]
profile = "black"
skip_glob = [
    "**/.fpllm/**",
    "**/.venv/**",
    "venv/**",
    "build/**",
    "dist/**",
    "**/.git/**",
    "**/.mypy_cache/**",
    "**/.pytest_cache/**",
    "**/.ruff_cache/**",
]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
